Extract repair_runner_contract_tests.rs from turn.rs (parent #680)
Move the two #[cfg(test)] mods that sit at the top of turn.rs into a new
src/agent/loop_run/repair_runner_contract_tests.rs sibling test file:

- repair_lifecycle_event_tests — pure-fn timeout / shadow-validation
  assertions on `repair_driver` + `verifier_orchestration` types.
- v0421_repair_runner_contract_tests — source-string grep assertions
  that pin production invariants in `turn.rs`,
  `verifier_orchestration.rs`, and `actor_loop_flow.rs`.

The `include_str!("turn.rs")` / `include_str!("verifier_orchestration.rs")`
/ `include_str!("actor_loop_flow.rs")` resolution paths are preserved
because the new sibling lives in the same `src/agent/loop_run/` directory.
Inner-mod paths updated from `super::super::X` to `super::super::X`
(unchanged after relocation since the parent layer is now the new file
itself; sibling crate paths reach `loop_run::*` via `super::super::*`).

Both mods stay `#[cfg(test)]` so they remain excluded from the production
binary. No facade re-export (DR3-001).

turn.rs: 26,644 -> 26,440 (-204 LOC)
Cumulative from 39,489: -13,049 LOC (-33.0%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
